292: move entire fsm to helpers/fsm.js
